Tilføjet logik der opretter visninger 3 måneder frem i tiden ved oprettelse af biograf og andre smårettelser

+++ src/main/java/com/example/kinobackend/order/OrderController.java
@@ -21,7 +21,7 @@ public class OrderController {
     }
 
     @GetMapping("/{orderId}")
-    public Order getOrderWithTickets(@PathVariable Long orderId) {
+    public Order getOrderWithTickets(@PathVariable int orderId) {
        return orderService.findOrderWithTickets(orderId);
    }
 
@@ -31,7 +31,7 @@ public class OrderController {
     }
 
     @GetMapping("/showing/{showingId}")
-    public List <Order> getOrderFromShowingId(@PathVariable Long showingId){
+    public List <Order> getOrderFromShowingId(@PathVariable int showingId){
         return orderService.findOrderFromShowingId(showingId);
     }
