+++ src/main/java/com/example/kinobackend/ticket/ITicketRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ITicketRepository extends JpaRepository<Ticket, Long> {
+public interface ITicketRepository extends JpaRepository<Ticket, Integer> {
     List<Ticket> findByOrder(Order order);
 }
